monitor-fastapi/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.0
SQLAlchemy==2.0.23
pydantic==1.10.10
python-jose==3.3.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.7
python-dotenv==1.0.0

